% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beijingdistricts}
\alias{Beijingdistricts}
\title{Boundaries of districts in Beijing}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 111 rows and 2 columns.
}
\usage{
Beijingdistricts
}
\description{
Boundaries of districts in Beijing
}
\keyword{datasets}
