% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depmunic}
\alias{depmunic}
\title{Municipality departments of Athens}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 8 columns.
}
\usage{
depmunic
}
\description{
Municipality departments of Athens
}
\details{
An sf object of 7 polygons with the following 7 variables:
\describe{
\item{num_dep}{An unique identifier for each municipality department.}
\item{airbnb}{The number of airbnb properties in 2017}
\item{museums}{The number of museums}
\item{population}{The population recorded in census at 2011.}
\item{pop_rest}{The number of citizens that the origin is a non european country.}
\item{greensp}{The area of green spaces (unit: square meters).}
\item{area}{The area of the polygon (unit: square kilometers). }
}
}
\keyword{datasets}
