% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\docType{data}
\name{flow}
\alias{flow}
\title{Flow data between residence and hospitals.}
\format{A data frame with 1178 rows and 2 variables:
\describe{
  \item{from}{home location}
  \item{to}{hospital}
  ...
}}
\usage{
flow
}
\description{
A dataset containing the home location and hospital visited by 1178 individuals (made up data).
}
\keyword{datasets}
