% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hsa_clust.R
\name{gen_hsa_clust}
\alias{gen_hsa_clust}
\title{Generate HSAs via Delamater's two step clustering approach}
\usage{
gen_hsa_clust(shp, from, to, li_threshold = 0.4,
  min_interventions = 10, fill = TRUE, min_clusters = 2,
  max_clusters = NULL, peaks = "dela")
}
\arguments{
\item{shp}{shapefile SpatialPolygonsDataFrame}

\item{from}{patients home region}

\item{to}{patients destination}

\item{li_threshold}{localization index threshold}

\item{min_interventions}{minimum number of interventions for HSAs to have}

\item{fill}{logical; fill in sources without any patients based on neighbours (defaults to TRUE) see details}

\item{min_clusters}{minimum number of HSAs in final solution}

\item{max_clusters}{maximum number of HSAs in final solution}
}
\value{
object of class hsa containing the following elements:
\describe{
  \item{\code{call}}{call to \code{gen_hsa}}
  \item{\code{lookup}}{final lookup table}
  \item{\code{original_data}}{list containing from and to originally submitted

    \code{from}: \code{from} as submitted to \code{gen_hsa}

    \code{to}: \code{to} as submitted to \code{gen_hsa}}
  \item{\code{reassigned_data}}{list containing from and to as assigned to HSAs

    \code{from}: \code{from} converted to its HSA

    \code{to}: \code{to} converted to its HSA}
  \item{\code{method}}{method used to assign each region to its HSA}
  \item{\code{original_shp}}{shapefile submitted to \code{gen_hsa} (via \code{shp})}
  \item{\code{shp}}{shapefile containing HSAs}
  \item{\code{iterations}}{number of iterations}
  \item{\code{li}}{localization index of the HSAs}
  \item{\code{li_threshold}}{\code{li_threshold} option}
  \item{\code{min_interventions}}{\code{min_interventions} option}
  \item{\code{n}}{number of HSAs created}
  \item{\code{n_it}}{number of HSAs in each iteration}
  \item{\code{n_interventions}}{number of interventions in each HSA after the final iteration}
  \item{\code{names}}{name of each HSA}
  \item{\code{flow}}{flow amongst HSAs (as generated by \code{flows})}

}
}
\description{
Generate HSAs via Delamater's two step clustering approach
}
\details{
Rownames of \code{shp} should be in common with \code{from} and \code{to}. If this is not the case, warnings (if some are missing) or errors will be issued (if no names are in common). Rownames can be set via, e.g., \code{spChFIDs(shp, shp@data$name)}.
         If \code{fill} is set to \code{TRUE}, observations from neighbouring regions are added to \code{from} to enable the missing region to be allocated to a HSA. The observations are not used in calculating the LI for the heuristic selection of the appropriate number of HSAs though.
}
