\name{HSAR-package}
\alias{HSAR-package}
\alias{HSAR}
\docType{package}
\title{
Hierarchical Spatial Autoregressive Model (HSAR)
}
\description{
A library of the Hierarchical Spatial Autoregressive Model (HSAR), based on a Bayesian Markov Chain Monte Carlo (MCMC) algorithm. }
\details{
\tabular{ll}{
Package: \tab HSAR\cr
Type: \tab Package\cr
Version: \tab 0.3.6\cr
Date: \tab 2015-12-21\cr
License: \tab GPL (>= 2)
}

}

\author{
Guanpeng Dong, Richard Harris, Angelos Mimis
Maintainer: Angelos Mimis <mimis@panteion.gr>
}

\references{
Dong, D. and Harris, R. 2014. Spatial Autoregressive Models for Geographically Hierarchical Data Structures. Geographical Analysis, 1-19. }

\keyword{ spatial econometrics }
