% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{The spatial locations of the Beijing land price data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1117 rows and 3 columns.
}
\usage{
land
}
\description{
The spatial locations of the Beijing land price data
}
\keyword{datasets}
