% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landprice}
\alias{landprice}
\title{Leased residential land parcels, from 2003 to 2009 in Beijing, China}
\format{
An object of class \code{data.frame} with 1117 rows and 11 columns.
}
\usage{
landprice
}
\description{
Leased residential land parcels, from 2003 to 2009 in Beijing, China
}
\details{
A \code{data.frame} with 1117 observations on the following 11 variables.
\describe{
\item{obs}{An unique identifier for each land parcel.}
\item{lnprice}{The log of the leasing price per square metre of each residential land parcel (unit: RMB, Chinese yuan) }
\item{dsubway}{The log of the distance of each land parcel to the nearest railway station (unit:meters)}
\item{dele}{The log of the distance of each land parcel to the nearest elementary school (unit:meters) }
\item{dpark}{The log of the distance of each land parcel to the nearest green park (unit:meters) }
\item{lnarea}{The log of the size of each land parcel (unit: square meters).}
\item{lndcbd}{The log of the distance of each land parcel to the CBD (centre business district) in Beijing (unit:meters) }
\item{year}{The year when each land parcel was leased with values of 0,1,2,3,4,5,6 representing year 2003,2004,2005,2006,2007,2008,2009}
\item{popden}{The population density of each district (unit: 1000 persons per square kilometers) }
\item{crimerate}{The number of reported serious crimes committed in each district per 1000 persons.}
\item{district.id}{The identifier of the district where each land parcel is located.}
}
}
\keyword{datasets}
