% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap.r
\name{minimap}
\alias{minimap}
\title{Minimaps of generated HSAs}
\usage{
minimap(object, ...)
}
\arguments{
\item{object}{object of hsa class (i.e. from gen_hsa(...))}

\item{hsa}{string - a regex expression defining which HSA(s) to plot}

\item{lake}{shapefile containing lakes (SpatialPolygons format)}

\item{col}{colour to use for focus area}

\item{bcol}{HSA border colour}

\item{regbcol}{colour to use for region borders}

\item{lcol}{colour to use for lakes}
}
\description{
Minimaps of generated HSAs
}
