% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\docType{data}
\name{shape}
\alias{shape}
\title{Shapefile for the flow dataset}
\format{A SpatialPolygonsDataFrame with 21 polygons and 2 variables:
\describe{
  \item{reg}{string, home location}
  \item{x}{numeric, random numbers}
  ...
}}
\usage{
shape
}
\description{
A shapefile containing the spatial data for the flow dataset.
}
\keyword{datasets}
