\name{summary.mcmc_hsar_lambda_0}
\alias{summary.mcmc_hsar_lambda_0}
\alias{print.mcmc_hsar_lambda_0}

\title{summary method for class mcmc_hsar_lambda_0}
\description{
 Methods for presenting the result of a fitted HSAR model with \eqn{\lambda} equal to 0. This is a model without the consideration of the possible interdependency between higher-level spatial units.
}
\usage{
  \method{summary}{mcmc_hsar_lambda_0}(object, ...)
  \method{print}{mcmc_hsar_lambda_0}(x, ...)
}

\arguments{
  \item{object}{\code{mcmc_hsar_lambda_0} An mcmc_hsar_lambda_0 object returned from the \code{hsar} function  }
  \item{x}{\code{mcmc_hsar_lambda_0} An mcmc_hsar_lambda_0 object returned from the \code{hsar} function  }
  \item{\dots}{Other arguments passed through}
}

\seealso{\code{\link{hsar}}}

\keyword{spatial econometrics}
