\name{summary.mcmc_hsar}
\alias{summary.mcmc_hsar}
\alias{print.mcmc_hsar}

\title{summary method for class mcmc_hsar}
\description{
  Methods for presenting the result of a fitted HSAR model.
}
\usage{
\method{summary}{mcmc_hsar}(object, ...)
\method{print}{mcmc_hsar}(x, ...)
}
  
\arguments{
  \item{object}{\code{mcmc_hsar} An mcmc_hsar object returned from the \code{hsar} function  }
  \item{x}{\code{mcmc_hsar} An mcmc_hsar object returned from the \code{hsar} function  }
  \item{\dots}{Other arguments passed through}
}

\seealso{\code{\link{hsar}} }

\keyword{spatial econometrics}
