\name{summary.mcmc_sar}
\alias{summary.mcmc_sar}
\alias{print.mcmc_sar}

\title{summary method for class mcmc_sar}
\description{
  Methods for presenting the result of a fitted standard spatial econometric model or a spatially lagged dependent variable model
}
\usage{
\method{summary}{mcmc_sar}(object, ...)
\method{print}{mcmc_sar}(x, ...)
}

\arguments{
  \item{object}{\code{mcmc_sar} An mcmc_sar object returned from the \code{sar} function  }
  \item{x}{\code{mcmc_sar} An mcmc_sar object returned from the \code{sar} function  }
  \item{\dots}{Other arguments passed through}
}

\seealso{ \code{\link{sar}} }

\keyword{spatial econometrics}